Stored NavUser as a component
